api/requirements.txt
flask
flask-restful
flask-cors
flask-sqlalchemy
psycopg
gunicorn
requests
pyopenssl
psycopg2
